Fixed build warning

+++ UnitTestCPP/unittestspage.cpp
@@ -120,7 +120,7 @@ void UnitTestsPage::Initialize(TestSummary* summary)
 
 void UnitTestsPage::Clear()
 {
-    for(size_t i = 0; i < m_dvListCtrlErrors->GetItemCount(); ++i) {
+    for(int i = 0; i < m_dvListCtrlErrors->GetItemCount(); ++i) {
         UTLineInfo* info = (UTLineInfo*)m_dvListCtrlErrors->GetItemData(m_dvListCtrlErrors->RowToItem(i));
         wxDELETE(info);
     }
